+++ hack/caicloud/sync-images.sh
@@ -25,7 +25,7 @@ KUBE_ROOT=$(dirname "${BASH_SOURCE}")/../..
 grep -IhEro "gcr.io/google_containers/[^\", ]*" \
      --include \*.go --include \*.json --include \*.yaml --include \*.yaml.in --include \*.yml --include Dockerfile --include \*.manifest \
      --exclude-dir=${KUBE_ROOT}/examples/guestbook \
-     ${KUBE_ROOT}/test ${KUBE_ROOT}/examples ${KUBE_ROOT}/cluster/addons ${KUBE_ROOT}/contrib ${KUBE_ROOT}/docs | sort -u |
+     ${KUBE_ROOT}/test ${KUBE_ROOT}/examples ${KUBE_ROOT}/cluster/addons ${KUBE_ROOT}/cluster/saltbase ${KUBE_ROOT}/contrib ${KUBE_ROOT}/docs | sort -u |
   while read -r gcr_image ; do
     image=${gcr_image#"gcr.io/google_containers/"}
     caicloudgcr_image="caicloudgcr/google_containers_$image"